requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==4.2.9
django-bootstrap-v5==1.0.11
django-filter==23.5
django-jazzmin==2.6.0
djangorestframework==3.14.0
Markdown==3.5.2
pillow==10.2.0
pytz==2024.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
